Lyric Search Functionality Completed

+++ spotify_data_scraper.py
@@ -4,6 +4,7 @@ import os
 from tkinter import Tk
 from tkinter.filedialog import askopenfilename
 import shutil
+import sys
 
 Tk().withdraw()
 spotify_directory = "./spotify_data/"
@@ -23,14 +24,15 @@ def get_song_data():
     for file in os.listdir(spotify_directory):
         if file.endswith(".json"):
             raw_data = read(spotify_directory + file)
-
             for entry in raw_data:
-                entry_data = [entry["trackName"], entry["artistName"]]
-                if entry_data not in songs:
-                    songs.append(entry_data)
-                else:
-                    songs.remove(entry_data)
-                    songs.append(entry_data)
+                if entry["trackName"] != "Unknown Track" or entry["artistName"] != "Unknown Artist":
+                    entry_data = [entry["trackName"], entry["artistName"]]
+                    if entry_data not in songs:
+                        songs.append(entry_data)
+                    else:
+                        songs.remove(entry_data)
+                        songs.append(entry_data)
+            print("IMPORTED: " + spotify_directory + file)
 
     songs.reverse()
 
@@ -45,38 +47,44 @@ def get_spotify_file(fresh=False):
         title="Select Spotify JSON",
         filetypes=(("JSON", "*.json"), ("All Files", "*.*")))
 
-    if not fresh:
+    if file_location != '':
+        if not fresh:
+            resolved_input = False
+
+            while not resolved_input:
+                remove_old = input(
+                    "Do you want to remove pre-existing files? [y/N]\n").lower()
+                if remove_old == "y":
+                    for file in os.listdir(spotify_directory):
+                        os.remove(spotify_directory + file)
+                    print("Files removed.\n")
+                    resolved_input = True
+                elif remove_old == "n" or not remove_old:
+                    print("Files not removed.\n")
+                    resolved_input = True
+                else:
+                    print("Unknown Input\n")
+
         resolved_input = False
 
         while not resolved_input:
-            remove_old = input(
-                "Do you want to remove pre-existing files? [y/N]\n").lower()
-            if remove_old == "y":
-                for file in os.listdir(spotify_directory):
-                    os.remove(spotify_directory + file)
-                print("Files removed.\n")
+            copy_or_move = input(
+                "Do you want to copy or move the selected file? [C/m]\n").lower()
+            if copy_or_move == "c" or not copy_or_move:
+                shutil.copy(file_location, spotify_directory)
+                print("Files Copied.\n")
                 resolved_input = True
-            elif remove_old == "n" or not remove_old:
-                print("Files not removed.\n")
+            elif copy_or_move == "m":
+                shutil.move(file_location, spotify_directory)
+                print("Files Moved.\n")
                 resolved_input = True
+
             else:
                 print("Unknown Input\n")
 
-    resolved_input = False
-
-    while not resolved_input:
-        copy_or_move = input(
-            "Do you want to copy or move the selected file? [C/m]\n").lower()
-        if copy_or_move == "c" or not copy_or_move:
-            shutil.copy(file_location, spotify_directory)
-            print("Files Copied.\n")
-            resolved_input = True
-        elif copy_or_move == "m":
-            shutil.move(file_location, spotify_directory)
-            print("Files Moved.\n")
-            resolved_input = True
-        else:
-            print("Unknown Input\n")
+        return True
+    else:
+        return False
 
 
 def data_present():
